chore: refactor ejected test-react-app to esm, real browser tests pass
