+++ packages/test-react-app/config/getHttpsConfig.cjs
@@ -4,7 +4,7 @@ const fs = require('fs');
 const path = require('path');
 const crypto = require('crypto');
 const chalk = require('react-dev-utils/chalk');
-const paths = require('./paths');
+const paths = require('./paths.cjs');
 
 // Ensure the certificate and key provided are valid and if not
 // throw an easy to debug error
