+++ packages/test-react-app/config/modules.cjs
@@ -2,7 +2,7 @@
 
 const fs = require('fs');
 const path = require('path');
-const paths = require('./paths');
+const paths = require('./paths.cjs');
 const chalk = require('react-dev-utils/chalk');
 const resolve = require('resolve');
