+++ packages/test-react-app/config/webpack/persistentCache/createEnvironmentHash.cjs
