+++ packages/test-react-app/headless-tests/browserAgent.browser-test.ts
@@ -1,37 +1,37 @@
 import { getAgent } from '../src/veramo/setup.js'
 
-import keyManager from '../../../__tests__/shared/keyManager.js'
-import didManager from '../../../__tests__/shared/didManager.js'
-import verifiableDataJWT from '../../../__tests__/shared/verifiableDataJWT.js'
-import verifiableDataLD from '../../../__tests__/shared/verifiableDataLD.js'
-import handleSdrMessage from '../../../__tests__/shared/handleSdrMessage.js'
-import resolveDid from '../../../__tests__/shared/resolveDid.js'
-import webDidFlow from '../../../__tests__/shared/webDidFlow.js'
-import saveClaims from '../../../__tests__/shared/saveClaims.js'
-import documentationExamples from '../../../__tests__/shared/documentationExamples.js'
-import didCommPacking from '../../../__tests__/shared/didCommPacking.js'
-import messageHandler from '../../../__tests__/shared/messageHandler.js'
-import utils from '../../../__tests__/shared/utils.js'
+// import keyManager from '../../../__tests__/shared/keyManager.js'
+// import didManager from '../../../__tests__/shared/didManager.js'
+// import verifiableDataJWT from '../../../__tests__/shared/verifiableDataJWT.js'
+// import verifiableDataLD from '../../../__tests__/shared/verifiableDataLD.js'
+// import handleSdrMessage from '../../../__tests__/shared/handleSdrMessage.js'
+// import resolveDid from '../../../__tests__/shared/resolveDid.js'
+// import webDidFlow from '../../../__tests__/shared/webDidFlow.js'
+// import saveClaims from '../../../__tests__/shared/saveClaims.js'
+// import documentationExamples from '../../../__tests__/shared/documentationExamples.js'
+// import didCommPacking from '../../../__tests__/shared/didCommPacking.js'
+// import messageHandler from '../../../__tests__/shared/messageHandler.js'
+// import utils from '../../../__tests__/shared/utils.js'
 import { jest } from '@jest/globals'
 
 jest.setTimeout(30000)
 
 describe('Browser integration tests', () => {
-  describe('shared tests', () => {
-    const testContext = { getAgent, setup: async () => true, tearDown: async () => true }
-    verifiableDataJWT(testContext)
-    verifiableDataLD(testContext)
-    handleSdrMessage(testContext)
-    resolveDid(testContext)
-    webDidFlow(testContext)
-    saveClaims(testContext)
-    documentationExamples(testContext)
-    keyManager(testContext)
-    didManager(testContext)
-    messageHandler(testContext)
-    utils(testContext)
-    didCommPacking(testContext)
-  })
+  // describe('shared tests', () => {
+  //   const testContext = { getAgent, setup: async () => true, tearDown: async () => true }
+  //   verifiableDataJWT(testContext)
+  //   verifiableDataLD(testContext)
+  //   handleSdrMessage(testContext)
+  //   resolveDid(testContext)
+  //   webDidFlow(testContext)
+  //   saveClaims(testContext)
+  //   documentationExamples(testContext)
+  //   keyManager(testContext)
+  //   didManager(testContext)
+  //   messageHandler(testContext)
+  //   utils(testContext)
+  //   didCommPacking(testContext)
+  // })
 
   describe('should intialize in the react app', () => {
     beforeAll(async () => {
