+++ packages/test-react-app/scripts/build.cjs
@@ -12,7 +12,7 @@ process.on('unhandledRejection', err => {
 });
 
 // Ensure environment variables are read.
-require('../config/env');
+require('../config/env.cjs');
 
 const path = require('path');
 const chalk = require('react-dev-utils/chalk');
@@ -20,7 +20,7 @@ const fs = require('fs-extra');
 const bfj = require('bfj');
 const webpack = require('webpack');
 const configFactory = require('../config/webpack.config');
-const paths = require('../config/paths');
+const paths = require('../config/paths.cjs');
 const checkRequiredFiles = require('react-dev-utils/checkRequiredFiles');
 const formatWebpackMessages = require('react-dev-utils/formatWebpackMessages');
 const printHostingInstructions = require('react-dev-utils/printHostingInstructions');
